prerequisites.sh: Don't fail if lsb_release executable missing
The first part of the script only warned if the /usr/bin/lsb_release
executable is missing, but the script later failed.  If lsb_release
is run successfully, then the release version has been already stored
in $release, so we can reuse that value to avoid running lsb_release
again (and if lsb_release is missing, this fix allows the script to
succeed anyway).  The change also changes to numerical comparison, to
implement the identified need to handle later versions beyond Ubuntu
18.04.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>